+++ lib/rapid-vaults/api.rb
@@ -15,8 +15,6 @@ class RapidVaults::API
       raise 'no file specified for encryption or decryption' if !settings.key?(:file)
     end
 
-    raise 'input password cannot be empty' if settings.key?(:pw) && settings[:pw].empty?
-
     # establish settings for api and denote using api
     settings.merge({ ui: :api })
   end