+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### 1.4.0 (Next)
 - Support specifying a custom basename for output files.
+- Improve password input validation.
 
 ### 1.3.2
 - Add `--force` CLI flag, and do not overwrite files by default.